feat: implement VisualizationManager for enhanced evaluation plotting capabilities
